reduce code duplication, add '-d' flag to xvfb-run
According to the tool's help page, the '-d' flag (auto-display) tells xvfb-run to "[…] find a display number automatically".
This fixes several integration tests which, when run locally, fail due to xvfb being unable to allocate a display number.
On Travis, this does not occur, because the test execution code checks for the presence of the "TRAVIS" environment variable.
(… Which should probably be refactored, too, removing the environment variable check in the business logic, in favor of a proper mocking solution in the test code!)